+++ src/log-collector/LogCollectorLaunchInfo.ts
@@ -1,4 +1,6 @@
 export interface LogCollectorLaunchInfo {
   commandName: string;
   projectDir: string;
+  shell: string;
+  root?: string;
 }