+++ src/log-collector/launchWithLogCollector.ts
@@ -1,8 +1,9 @@
 import { startShellCommand } from '@facetlayer/subprocess-wrapper';
 import * as Path from 'node:path';
 import { ProjectRootDir } from '../dirs.ts';
+import type { LogCollectorLaunchInfo } from './LogCollectorLaunchInfo.ts';
 
-export async function launchWithLogCollector(commandName: string, projectDir: string) {
+export async function launchWithLogCollector(launchInfo: LogCollectorLaunchInfo) {
   const command = process.argv[0];
   const args = [Path.join(ProjectRootDir, 'dist', 'main-log-collector.js')];
 
@@ -19,12 +20,7 @@ export async function launchWithLogCollector(commandName: string, projectDir: st
 
   await subprocess.waitForStart();
 
-  subprocess.proc.stdin.write(
-    JSON.stringify({
-      commandName,
-      projectDir,
-    })
-  );
+  subprocess.proc.stdin.write(JSON.stringify(launchInfo));
 
   subprocess.proc.stdin.end();
   return subprocess;